Change celestial map
Modified the celestial GUI selection screen to allow for other solar
systems to be used properly. Also changed the screens to only show the
main solar system. This isn't the best implementation of this, although
it's a basic way to implement it so addons can be used to their full
extent. I imagine more optimisation, feature additions, and better
implementations of this will be added to do a better job than what I
did :)